restructuring code to separate tests and src
