+++ DataFrame/include/DataFrame/DataFrameOperators.h
@@ -8,6 +8,8 @@
 #ifndef HMDF_DATAFRAMEOPERATORS_HPP
 #define HMDF_DATAFRAMEOPERATORS_HPP
 
+#include "DataFrame_lib_exports.h"
+
 #include "DataFrame.h"
 // #include <execution>
