+++ DataFrame/include/DataFrame/HeteroVector.h
@@ -1,49 +1,83 @@
 // Hossein Moein
-// October 24, 2018
+// September 11, 2017
 // Copyright (C) 2018-2019 Hossein Moein
 // Distributed under the BSD Software License (see file License)
 
 #pragma once
 
-#ifndef HMDF_HETEROVIEW_HPP
-#define HMDF_HETEROVIEW_HPP
+#ifndef HMDF_HETEROVECTOR_HPP
+#define HMDF_HETEROVECTOR_HPP
 
-#include "VectorView.h"
+#include "DataFrame_lib_exports.h"
+
+#include "HeteroView.h"
+#include <vector>
 #include <unordered_map>
-#include <functional>
 
 // ----------------------------------------------------------------------------
 
 namespace hmdf
 {
 
-struct  HeteroView  {
+// This class implements a heterogeneous vector. Its design and implementation
+// are partly inspired by Andy G's Blog at:
+// https://gieseanw.wordpress.com/2017/05/03/a-true-heterogeneous-container/
+//
+struct  HeteroVector  {
 
 public:
 
     using size_type = size_t;
 
-    HeteroView() = default;
-    template<typename T>
-    HeteroView(T *begin_ptr, T *end_ptr);
-    HeteroView(const HeteroView &that);
-    HeteroView(HeteroView &&that);
+    HeteroVector();
+    HeteroVector(const HeteroVector &that);
+    HeteroVector(HeteroVector &&that);
 
-    ~HeteroView() { clear(); }
+    ~HeteroVector() { clear(); }
 
-    HeteroView &operator= (const HeteroView &rhs);
-    HeteroView &operator= (HeteroView &&rhs);
+    HeteroVector &operator= (const HeteroVector &rhs);
+    HeteroVector &operator= (HeteroVector &&rhs);
 
     template<typename T>
-    VectorView<T> &get_vector();
+    std::vector<T> &get_vector();
     template<typename T>
-    const VectorView<T> &get_vector() const;
+    const std::vector<T> &get_vector() const;
+
+    // It returns a view of the underlying vector.
+    // NOTE: One can modify the vector through the view. But the vector
+    //       cannot be extended or shrunk through the view.
+    // There is no const version of this method
+    //
+    template<typename T>
+    HeteroView get_view(size_type begin = 0, size_type end = -1);
 
     template<typename T>
-    typename VectorView<T>::
+    void push_back(const T &v);
+    template<typename T, class... Args>
+    void emplace_back (Args &&... args);
+    template<typename T, typename ITR, class... Args>
+    void emplace (ITR pos, Args &&... args);
+
+    template<typename T>
+    void reserve (size_type r)  { get_vector<T>().reserve (r); }
+    template<typename T>
+    void shrink_to_fit () { get_vector<T>().shrink_to_fit (); }
+
+    template<typename T>
     size_type size () const { return (get_vector<T>().size()); }
 
     void clear();
+
+    template<typename T>
+    void erase(size_type pos);
+
+    template<typename T>
+    void resize(size_type count);
+    template<typename T>
+    void resize(size_type count, const T &v);
+
+    template<typename T>
+    void pop_back();
 
     template<typename T>
     bool empty() const noexcept;
@@ -66,14 +100,13 @@ public:
 private:
 
     template<typename T>
-    static std::unordered_map<const HeteroView *, VectorView<T>>    views_;
+    static std::unordered_map<const HeteroVector *, std::vector<T>> vectors_;
 
-    std::function<void(HeteroView &)>   clear_function_ {
-        [](HeteroView &) { return; } };
-    std::function<void(const HeteroView &, HeteroView &)>   copy_function_ {
-        [](const HeteroView &, HeteroView &)  { return; } };
-    std::function<void(HeteroView &, HeteroView &)> move_function_  {
-        [](HeteroView &, HeteroView &)  { return; } };
+    std::vector<std::function<void(HeteroVector &)>>    clear_functions_;
+    std::vector<std::function<void(const HeteroVector &,
+                                   HeteroVector &)>>    copy_functions_;
+    std::vector<std::function<void(HeteroVector &,
+                                   HeteroVector &)>>    move_functions_;
 
     // Visitor stuff
     //
@@ -145,10 +178,10 @@ public:
 // ----------------------------------------------------------------------------
 
 #  ifndef DMS_NOINCLUDE_SOURCE
-#    include "HeteroView.tcc"
+#    include "HeteroVector.tcc"
 #  endif // DMS_NOINCLUDE_SOURCE
 
-#endif //HMDF_HETEROVIEW_HPP
+#endif //HMDF_HETEROVECTOR_HPP
 
 // ----------------------------------------------------------------------------
 
@@ -157,4 +190,3 @@ public:
 // tab-width:4
 // c-basic-offset:4
 // End:
-
