+++ test/vectors_tester.cc
@@ -3,7 +3,7 @@
 #include <string>
 #include <typeinfo>
 
-#include "../include/HeteroVector.h"
+#include "../DataFrame/include/DataFrame/HeteroVector.h"
 
 using namespace hmdf;
